requirements.txt
asgiref
django
gunicorn
whitenoise
requests
urllib3
psycopg2==2.9.9
psycopg2-binary
dj-database-url
sqlparse
tzdata
djangorestframework
djangorestframework-simplejwt
markdown
django-filter
pillow
django-cors-headers